* numbers.c (scm_gcd): change "k" to a long from an int. Otherwise it fails on the alpha.  However, we might rather choose this size conditionally.

+++ libguile/numbers.c
@@ -383,7 +383,7 @@ scm_gcd (SCM x, SCM y)
       } else if (yy == 0) {
 	result = u;
       } else {
-	int k = 1;
+	long k = 1;
 	long t;
 
 	/* Determine a common factor 2^k */
